chore: create configuration layer
Move configuration logic inside of layer
